Extracts song name into separate component
