+++ src/modules/songs/components/index.ts
@@ -2,3 +2,4 @@ export * from "./SongDetailsLabel/SongDetailsLabel";
 export * from "./SongDetailsItem/SongDetailsItem";
 export * from "./SongDetailsPage/SongDetailsPageLoading/SongDetailsPageLoading";
 export * from "./SongDetailsPage/SongDetailsPageHeader/SongDetailsPageHeader";
+export * from "./SongDetailsPage/SongDetailsPageSongName/SongDetailsPageSongName";
